requirements.txt
kafka-python
pyyaml
aiokafka